Added project path support
Implemented GetProjectPath and SetProjectPath functions in the utils module to manage the FS_PROJECT_PATH environment variable cross-functionally between different operating systems. The SetProjectPath function adjusts the variable for Windows and Unix-like systems, handling its update according to the OS, in addition to providing feedback in case of failures in the process.